refactor: split markdown ordering helpers
